Fix naming convention of interface classes
