+++ xsd-core/src/main/java/io/ballerina/xsdtorecordconverter/XSDFactory.java
@@ -20,19 +20,19 @@ package io.ballerina.xsdtorecordconverter;
 
 import io.ballerina.xsdtorecordconverter.component.ComplexType;
 import io.ballerina.xsdtorecordconverter.component.Element;
-import io.ballerina.xsdtorecordconverter.component.IComponent;
+import io.ballerina.xsdtorecordconverter.component.XSDComponent;
 import io.ballerina.xsdtorecordconverter.component.SimpleType;
 import org.w3c.dom.Node;
 
-import static io.ballerina.xsdtorecordconverter.visitor.XSDVisitor.COMPLEX_TYPE;
-import static io.ballerina.xsdtorecordconverter.visitor.XSDVisitor.ELEMENT;
-import static io.ballerina.xsdtorecordconverter.visitor.XSDVisitor.SIMPLE_TYPE;
+import static io.ballerina.xsdtorecordconverter.visitor.XSDVisitorImpl.COMPLEX_TYPE;
+import static io.ballerina.xsdtorecordconverter.visitor.XSDVisitorImpl.ELEMENT;
+import static io.ballerina.xsdtorecordconverter.visitor.XSDVisitorImpl.SIMPLE_TYPE;
 
 /**
  * The factory class for generating XSD components.
  */
 public final class XSDFactory {
-    public static IComponent generateComponents(Node node) {
+    public static XSDComponent generateComponents(Node node) {
         if (node.getNodeType() != Node.ELEMENT_NODE) {
             return null;
         }
